+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Render search results as normal and show dash for empty status. Refs UIREQ-818.
 * Support inventory 12.0 in okapiInterfaces. Refs UIREQ-819.
 * Refactor forms to use final-form. Refs UIREQ-419.
+* Requests: Implement App context menu and keyboard shortcuts modal. refs UIREQ-817.
 
 ## [7.1.3](https://github.com/folio-org/ui-requests/tree/v7.1.3) (2022-08-11)
 [Full Changelog](https://github.com/folio-org/ui-requests/compare/v7.1.2...v7.1.3)